new_requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==2.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.2
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.7
